feat(ee): move core billing to ee/billing, canonical isBillingEnabled in ee/license
